api/pyproject.toml
[project]
name = "course-api"
version = "0.1.0"
description = "Simple wrapper around explorecourses library"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "explorecourses>=1.0.6",
    "fastapi>=0.104.0",
    "uvicorn[standard]>=0.24.0",
    "python-dotenv>=1.0.1",
]
